% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_unique.R
\name{check.unique}
\alias{check.unique}
\title{Small function to check for values in data.table based on column.
Mainly used to make sure the sample_id is actually unique.}
\usage{
check.unique(dat, columns = "sample_id")
}
\description{
Small function to check for values in data.table based on column.
Mainly used to make sure the sample_id is actually unique.
}

