% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_with_sql.R
\name{merge_with_sql}
\alias{merge_with_sql}
\title{Merge data table into PostgreSQL database}
\usage{
merge_with_sql(input, tablename, key = "code", dbname = "leaf_spectra", ...)
}
\description{
Merge data table into PostgreSQL database
}
\examples{
library(specprocess)
tablename <- 'projects'
input <- data.table(
   code = 'test',
   description = 'example')
key <- 'code'
dbname <- 'leaf_spectra'
merge_with_sql(input, tablename, key = 'code')
}

