% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlmat2list.R
\name{wlmat2list}
\alias{wlmat2list}
\title{Convert multi-column matrix where wavelength is first column to named list}
\usage{
wlmat2list(wlmat)
}
\description{
Convert multi-column matrix where wavelength is first column to named list
}

