% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_with_sql.R
\name{merge_with_sql}
\alias{merge_with_sql}
\title{Merge data table into PostgreSQL database}
\usage{
merge_with_sql(input, tablename, dbname = "leaf_spectra", by = NULL,
  verbose = FALSE)
}
\description{
Merge data table into PostgreSQL database
}

